course info, Separate Module
